rename: homeScreen -> home_screen 파일 이름 snake_case로 변경
# homeScreenfileRename
