+++ lib/routers/routers.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:swm_peech_flutter/features/home/view/homeScreen.dart';
+import 'package:swm_peech_flutter/features/home/view/home_screen.dart';
 import 'package:swm_peech_flutter/features/scriptInput/controller/script_input_controller.dart';
 import 'package:swm_peech_flutter/features/scriptInput/view/script_input_screen.dart';
 import 'package:swm_peech_flutter/features/themeInput/controller/theme_input_controller.dart';
